Fix order of work/predicate in function call
This ordering had changed during development and this was left over from
an earlier time when `predicate` preceded `work`.  The order was swapped
to put `work` first so that it could be implemented as a `do` block to
the `launch` function.  It is expected that `work` will be more
heavyweight than predicate, so it seemed more likely that the caller
would implement it as `do` block.